Move vpn and relay schemas to schemas/waitlist.py
